+++ matchbox-server/src/main/java/ch/ahdis/fhir/hapi/jpa/validation/MatchboxImplementationGuideProvider.java
@@ -10,4 +10,14 @@ import ca.uhn.fhir.jpa.packages.PackageInstallOutcomeJson;
 public interface MatchboxImplementationGuideProvider {
 
 	PackageInstallOutcomeJson loadAll(boolean replace);
+
+	/**
+	 * Returns whether the given ImplementationGuide is installed or not.
+	 */
+	boolean has(final String packageId, final String packageVersion);
+
+	/**
+	 * Installs the given ImplementationGuide from the internet registry.
+	 */
+	void installFromInternetRegistry(final String packageId, final String packageVersion);
 }