+++ docs/changelog.md
@@ -1,6 +1,8 @@
 Unreleased
 
 - Allow providing map and models in the StructureMap $transform operation [#305](https://github.com/ahdis/matchbox/issues/305)
+- Introduce parameter 'autoInstallMissingIgs' to automatically install IGs from the public registry
+  [#306](https://github.com/ahdis/matchbox/issues/306)
 
 2024/11/13 Release 3.9.7